refactor(executions): address PR review — move JsonTree to design system, add as const
- Move JsonTree.vue → KsJsonTree.vue in design-system Data components
- Register KsJsonTree globally in index.ts (import, components map, export, GlobalComponents)
- Update LogLine.vue to use KsJsonTree (globally registered, no local import)
- Move Storybook story to design-system tests/storybook/Data/KsJsonTree.stories.ts
- Delete old story from ui/tests
- Add as const to DENSITY_PADDING in useLogDisplay.ts
- Add as const to FILTERABLE_LOG_FIELDS in logValueFilter.ts; cast field key to keyof typeof
