+++ ui/src/composables/useLogDisplay.ts
@@ -2,11 +2,11 @@ import {useStorage} from "@vueuse/core"
 
 export type LogDensity = "compact" | "normal" | "expanded";
 
-export const DENSITY_PADDING: Record<LogDensity, string> = {
+export const DENSITY_PADDING = {
     compact: "2px",
     normal: "5px",
     expanded: "12px",
-}
+} as const
 
 export const logsFontSize = useStorage<number>("logsFontSize", 14)
 export const logsDensity = useStorage<LogDensity>("logsDensity", "normal")
